Document changes. Separate files. Enable new code.
